@@ -1,5 +1,3 @@
 namespace Frequency.Session;
 
-public class ObjectSession {
-
-}
+public class ObjectSession { }
