@@ -1,5 +1,3 @@
 namespace Frequency.Session.State;
 
-public class TerminalState {
-
-}
+public class TerminalState { }
