@@ -1,5 +1,3 @@
 namespace Frequency.Session;
 
-public class TerminalSession {
-
-}
+public class TerminalSession { }
